Expanded Movement
Expanded the movement using Input Actions. Added easing in and out for movement using Lerp (linear).

Known issues:
When both a controller and a keyboard is connected they will switch from time to time. (This may only occur on gamepads with gyroscopes that continuously give inputs)

When opposite movement keys are pressed the character halts to a stand still. This could be left as is or could be changed to having the last button pressed stay active.